==> code/02_Data_Cleaning-FMA.ipynb <==
# Data Cleaning

## Numeric Feature Data

# Imports
import pandas as pd

# Reading in the data
genre = pd.read_csv('../fma_data/genre.csv', skiprows=1)

# Taking a look at the data
genre.head(15)

### Creating Labels

# Fixing the file names and labels
genre['files'] = genre['files'].map(lambda x: x[11:-2])
genre['labels'] = genre['labels'].map(lambda x: x[11:])

# Mapping the labels to numeric values
label_map = {
    'Electronic': 0,
    'Experimental': 1,
    'Folk': 2,
    'Hip-Hop': 3,
    'Instrumental': 4,
    'International': 5,
    'Pop': 6,
    'Rock': 7
}

genre['y'] = genre['labels'].map(label_map)

genre.head()

#### Export

genre.to_csv('../fma_data/genre_clean.csv', index=False)

## Mel Spectrogram Data

# Reading in the data
mel_specs = pd.read_csv('../fma_data/genre_mel_specs.csv')

# Taking a look at the data
mel_specs.head()

### Creating Labels

# Renaming the label column and mapping them to numeric values using the same map as above
mel_specs = mel_specs.rename(columns={'84480': 'y'})
mel_specs['y'] = mel_specs['y'].map(label_map)

#### Export

mel_specs.to_csv('../fma_data/genre_mel_specs_clean.csv', index=False)
