==> code/99_Work_In_Progress.ipynb <==
# Modeling with Numeric Features

#### Important Note!
This notebook was temporarily put to the side because I decided to use a CNN with the mel specgrograms as features instead. I would like to return to this notebook on a later date and do some more research on numeric features to see how well they do.

# Imports
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier

# Reading in the data
genre = pd.read_csv('../data/genre_clean.csv')

# First five rows of the data for reference
genre.head()

# Defining our features and predictor variable
X = genre.drop(['files', 'labels', 'y'], axis=1)
y = genre['y']

# Getting our baseline accuracy
y.value_counts(normalize=True)

# Train test split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42, stratify=y)

# Scaling the features
ss = StandardScaler()
X_train_sc = ss.fit_transform(X_train)
X_test_sc = ss.transform(X_test)

## Support Vector Machine

# Instantiating and fitting a basic svc model
svc = SVC()
svc.fit(X_train_sc, y_train)

# Checking the accuracy
print(f'Score on the training set: {svc.score(X_train_sc, y_train)}')
print(f'Score on the test set: {svc.score(X_test_sc, y_test)}')

## Random Forest

# Instantiating and fitting a basic random forest model
rf = RandomForestClassifier()
rf.fit(X_train, y_train)

# Checking the accuracy
print(f'Score on the training set: {rf.score(X_train, y_train)}')
print(f'Score on the test set: {rf.score(X_test, y_test)}')
